+++ src/main/java/core/basesyntax/IsoscelesTrapezoid.java
@@ -1,16 +1,15 @@
 package core.basesyntax;
 
 public class IsoscelesTrapezoid extends Figure {
-    private double baseSide;
-    private double topSide;
-    private double height;
+    private int baseSide;
+    private int topSide;
+    private int height;
 
-    public IsoscelesTrapezoid(String color, double baseSide, double topSide, double height) {
+    public IsoscelesTrapezoid(String color, int baseSide, int topSide, int height) {
         this.height = height;
         this.topSide = topSide;
         this.baseSide = baseSide;
     }
-
 
     @Override
     public double calculateArea() {
@@ -20,6 +19,7 @@ public class IsoscelesTrapezoid extends Figure {
     @Override
     public String drawTheFigure() {
         return "Figure : isoscelesTrapezoid, area: " + calculateArea()
-                + " sq.units, sideA: " + baseSide + " units, sideB: " + topSide + " units, color: " + color.toLowerCase();
+                + " sq.units, sideA: " + baseSide + " units, sideB: "
+                + topSide + " units, color: " + color.toLowerCase();
     }
 }
