+++ src/main/java/core/basesyntax/RightTriangle.java
@@ -1,10 +1,10 @@
 package core.basesyntax;
 
 public class RightTriangle extends Figure {
-    private double firstLeg;
-    private double secondLeg;
+    private int firstLeg;
+    private int secondLeg;
 
-    public RightTriangle(String color, double firstLeg, double secondLeg) {
+    public RightTriangle(String color, int firstLeg, int secondLeg) {
         this.firstLeg = firstLeg;
         this.secondLeg = secondLeg;
     }
@@ -17,6 +17,7 @@ public class RightTriangle extends Figure {
     @Override
     public String drawTheFigure() {
         return "Figure: triangle, area: " + calculateArea() + " sq.units, firstLeg: "
-                + firstLeg + " units, secondLeg: " + secondLeg + " units, color: " + color.toLowerCase();
+                + firstLeg + " units, secondLeg: "
+                + secondLeg + " units, color: " + color.toLowerCase();
     }
 }
